feed_service/pyproject.toml
[project]
name = "feed-service"
version = "0.1.0"
description = "ASTU Connect Feed Service"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-jose[cryptography]",
    "sqlalchemy",
    "psycopg2-binary",
    "pydantic-settings",
    "aiokafka",
    "redis",
    "httpx",
]
